move models into their own files and also cleaned up anything a bit
